add an implementation of HTTPS Everywhere
A ruby script converts all of the XML rule files into hashes and
dumps them to plist files, which are parsed by the app at startup.

When intercepting URLs, check for matching rules by hostname (using
the same wildcarding as the Javascript in the Firefox extension) and
then rewrite them.  Supports exclusions as well.

Seems to be pretty fast, and has a cache size of (currently) 20
rules to avoid having to recompile all those regexes and stuff on
every request.