improve security of opening tabs to block spam
In window.open(), generate a hidden <a> and use a fake MouseEvent to
click on it.  This will cause the request to the UIWebView to come
with a UIWebViewNavigationType of UIWebViewNavigationTypeLinkClicked
just like if it were an <a target=_blank> and the user clicked on
it.

Trying to call our window.open() implementation from an onLoad or
similar will fail because the dispatchEvent() call will not do
anything in WebKit, meaning the UIWebView delegate won't see it and
the window.open() will fail.  We can cope with this failure better
in the future by showing a "popup blocked" message and allowing the
user to override it.