add functionality to block internet pages from accessing LAN hosts
Inspired by <https://code.google.com/p/chromium/issues/detail?id=378566>,
this adds a setting which defaults to on that will check each origin
host's DNS records and remember if any of them resolve to an RFC3330
local network IP.  If not, meaning the page is on the internet, then
block any page sub-requests that try to load from hosts that do
resolve to local networks.

This functionality is enough to block a silly attack like a webpage
on the internet having:

    <img src="http://192.168.100.1/reset.htm?reset_modem=Restart+Cable+Modem">

which does in fact restart my cable modem with this setting disabled
(and in Firefox).

Since iOS has its own internal DNS cache, these extra DNS lookups
do not actually hit the wire as verified with tcpdump (although it
would be nice to do this ourselves once and hand off an IP to the
URL loading system so we know for sure what it's talking to).