define a custom long-press context menu, add 'open in a new tab'
When long-pressing on an image, add function to download image and
save to the device.  If the image has alt text, show that in the
context menu (useful for sites like xkcd that have hidden text that
would otherwise require hovering over a link).

Tap coordinate translation and element data extraction concepts
borrowed from FoxBrowser.

Closes #4.