CKHTTPConnection: work around random "Frame load interrupted" failures
CFHTTPMessageIsHeaderComplete apparently lies on random occasions
and says that it has parsed complete header data for the response
when in fact it hasn't.  This resulted in passing up a response to
URLInterceptor's didReceiveResponse method where the response would
have a status of 200, a content-type of "application/octet-stream"
and 0 length headers.  Since a lot of these failures happened with
gzipped responses, the garbage data would try to be passed directly
through to WebView and fail loading with a "Frame load interrupted"
popup error.

To work around this, keep waiting for new data when the headers
parsed have a 0 length.