CookieJar: fix local storage cleaning
The regex change in 42642850300da changed the capture group that
-localStorageFiles was relying on to pluck the hostname out of the
filename.  Repair that and set a new define up near the regex to
make it clear that the two need to be in sync.

Walk the Cache directory recursively in -localStorageFiles and put
all non-host and non-excluded files into a "(Other cache data)"
pseudo-host that can be cleared by the user or when
-clearAllNonWhitelistedLocalStorageOlderThan:0 is run.
There was a ton of data in here on my phone, such as a fsCachedData
directory with a huge number of files containing cached HTML.

closes #40