WVT: disable ContentInsetAdjustmentBehavior
Apparently something in iOS changed recently which caused builds on
the iPhone X to always showing the UIWebView offset from the top of
the view, even though its frame was set properly.  Once the screen
was redrawn, such as by tapping the settings icon, the UIWebView
would jump back up to the top where it belonged.

Disabling ContentInsetAdjustmentBehavior seems to fix this when
toolbarOnBottom is on, and as a bonus, eliminates the small gray bar
at the bottom of a long page when toolbarOnBottom is off, allowing a
page to take up the entire screen like it does in Safari.