AppDelegate: set AVAudioSessionCategoryAmbient
This controls how iOS handles audio from media playing in UIWebView
such as videos.  By setting AVAudioSessionCategoryAmbient, iOS will
continue playing background audio from other apps when audio/video
starts playing in a UIWebView.

This means when playing audio in the background from another app
like a music or podcast app, viewing a video-style gif in Endless
won't pause that background audio.

This is how Safari behaves but it goes a step further and when
viewing a video that has sound, it will pause the background audio.
I'm not sure how to do that, but this will at least stop the
annoying behavior of stopping my podcast playing when viewing
puppers on gfycat.com.