WVC: fix disappearing keyboard in iOS 11.4
iOS 11.4 started sending additional UIKeyboardWillShowNotification
notifications when the user taps on the URL bar when the keyboard is
already up, so keyboardHeight ends up being 0 (because there is no
change in FrameBegin->FrameEnd) and disappears.

Keep track of whether the keyboard is showing and ignore these
duplicate events until we get a UIKeyboardWillHideNotification
event.