WVC: subtract keyboardHeight from root view height
If the software keyboard is on-screen, the web view can't show
underneath it or else it won't be able to scroll down far enough to
show the bottom of the page.

This is especially problematic if there are text fields near the
bottom since you won't be able to see the input you're typing.