WVC: fix layout issue after focusing url bar then viewing tabs
When viewDidLayoutSubviews is called during transition from keyboard
up to back down, self.view.safeAreaInsets.bottom is 0, which
viewDidLayoutSubviews uses to place the toolbar at the very bottom
of the screen.

Once the keyboard is actually down, viewSafeAreaInsetsDidChange is
then called slightly after which adjusts safeAreaInsets.bottom to
something bigger.

viewSafeAreaInsetsDidChange can't just call viewDidLayoutSubviews
though, because showingTabs is still true.  So work around this by
caching any non-zero safeAreaInsets.bottom and use it when the
keyboard is still up but transitioning down (when it would be zero).