Add a setting to adjust mute switch behavior
The previous behavior was to honor the mute switch to mute audio
played through web pages and video, so that a web page could never
play audio through the speakers unless you wanted it to by adjusting
the mute switch.

Since this behavior might be confusing to people (like the person
that left a 1-star review because the app didn't play sound), add an
option to disable this behavior and always play audio regardless of
the mute switch.